backend/requirements.txt
aerich==0.7.1
aiosqlite==0.17.0
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
celery==5.2.7
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.86.0
fastapi-jwt-auth==0.5.0
h11==0.12.0
httpx==0.23.0
idna==3.4
iso8601==1.1.0
passlib==1.7.4
pydantic==1.10.1
PyJWT==1.7.1
pypika-tortoise==0.1.6
pytz==2022.6
sniffio==1.3.0
starlette==0.20.4
tortoise-orm==0.19.2
typing_extensions==4.4.0
uvicorn==0.19.0
redis==4.3.4